+++ protocol/triple/triple-tool/gen/generator/genTriple.go
