+++ protocol/triple/triple-tool/gen/generator/tripleTpl.go
