+++ protocol/triple/triple-tool/gen/generator/tripleTpl_test.go
