+++ protocol/triple/triple-tool/main.go
@@ -15,37 +15,12 @@
  * limitations under the License.
  */
 
-package cmd
+package main
 
 import (
-	"os"
+	"dubbo.apache.org/dubbo-go/v3/triple-tool/cmd"
 )
 
-import (
-	"github.com/spf13/cobra"
-)
-
-import (
-	"dubbo.apache.org/dubbo-go/v3/protoc-gen-triple/gen"
-	"dubbo.apache.org/dubbo-go/v3/protoc-gen-triple/internal/version"
-)
-
-// rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
-	Use: "protoc-gen-triple",
-}
-
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
-	err := rootCmd.Execute()
-	if err != nil {
-		os.Exit(1)
-	}
-}
-
-func init() {
-	rootCmd.Version = version.Version
-
-	rootCmd.AddCommand(gen.Cmd)
+func main() {
+	cmd.Execute()
 }
